delete redundant parse_logs.py, add easy script for login to UPMEM machines and update memlimit structure
