+++ bench/setup-UPMEM.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+
+sudo apt-get install sysstat bc
+
+mkdir /scratch/db
+chmod 777 /scratch/db
